+++ docs/source/index.rst
@@ -12,7 +12,6 @@ The :mod:`torchaudio` package consists of I/O, popular datasets and common audio
    functional
    transforms
    datasets
-   metrics
    models
    sox_effects
    compliance.kaldi
