--- docs/source/metrics.rst
@@ -1,14 +1,0 @@
-.. role:: hidden
-    :class: hidden-section
-
-torchaudio.metrics
-======================
-
-.. currentmodule:: torchaudio.metrics
-
-The models subpackage contains definitions of metrics for addressing common audio tasks.
-
-:hidden:`lenvenshtein_distance`
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: levenshtein_distance
