+++ torchaudio/__init__.py
@@ -4,7 +4,6 @@ from torchaudio import (
     compliance,
     datasets,
     kaldi_io,
-    metrics,
     utils,
     sox_effects,
     transforms
